producer/requirements.txt
kafka-python==2.0.2
faker==30.3.0

